requirements.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
Django==5.0
django-filter==23.5
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
Markdown==3.5.1
mypy==1.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
pytz==2023.3.post1
rest-condition==1.0.3
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
